Retry Menu Update
Added a "retry menu" for after the player collides with a meteor, this is the main change.
Refactored temp_grass_tile_0.png into a new tiles directory within the sprites directory.
Updated project_notes.
Changed the game loop structure quite a bit with the retry menu addition. It is the best way I can think to go about structuring the game loop.
All that is needed now is to complete all the sprites/images and make some minor changes to the code. One example of how I will change the code is creating an overarching "Menu" class rather than have the main- and retry menu have their own classes.