Background Integration Update
Spent a lot of time creating a Background module that creates a randomly generated background based on a given tile-set. I hope to clean it up a bit and release it in a standalone repository.
Somewhat unrelated, but I also hope to create a demonstration/module of the "wave collapse" algorithm using the aforementioned Background module.